+++ entrypoints/background.ts
@@ -47,12 +47,18 @@ export default defineBackground(() => {
             esm = checkExports(data.exports);
           }
 
+          // Count dependencies (dependencies + peerDependencies)
+          const depCount =
+            Object.keys(data.dependencies || {}).length +
+            Object.keys(data.peerDependencies || {}).length;
+
           sendResponse({
             ok: true,
             data: {
               types,
               dtTypes,
               esm,
+              depCount,
             },
           });
         })